requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.31.0
flask==2.3.3
python-dotenv==1.0.0
gunicorn==21.2.0
schedule==1.2.2
